Issue32: "make install" is now optional. Users can now simply "configure", "make", and then add the "pylith3d" subdirectory of the source/build directory to their PATH.
In this mode, changes to C++ simply require a "make" from the
top-level directory. Changes to the Python code do not require you to
run "make" at all -- simply re-run 'pylith3dapp.py'.

Also: Changed the 'configure' option '--with-embedding' to
'--enable-embedding', for consistency with CitcomS.

Note: when configured with '--disable-embedding', "make install" is
NOT optional.  This is because the PyLith extension module has the
same name as the top-level package ('pylith3d').  The extension module
must be installed under the package directory in order for everything
to work.  There are two possible fixes: 1) tweak the build so that
'pylith3dmodule.so' lands under 'pylith3d'; or 2) rename the extension
module.