Tied PyLith's configuration even more closely to PETSc's: use 'petscconf' to determine both CXX and the MPI flags.
I'm not sure why this 'configure' script didn't grab MPI configuration
from PETSc.  But, I remember why it didn't grab CXX from PETSc:
Lithomop didn't require PETSc to be configured with C++, so it
couldn't assume that 'petscconf' defined CXX.  Now, with the addition
of the Sieve dependency, it can be assumed that 'petscconf' defines
CXX.

It is almost certainly an error to configure PyLith with a C++
compiler different from the one used to configure PETSc; this change
avoids that potential mistake.  This change also avoids confusing
'configure' error messages about MPIINCLUDES and MPILIBS, which aren't
relevant to PyLith -- PyLith should simply be configured with the MPI
library defined by the PETSc configuration.