Added a "keyval codec" so that *.keyval files can be processed by the Pyre framework.  Keyval parameters are now ordinary Inventory items, so they can also be set from the command line or a .cfg file.  Another advantage of this is error reporting: try inserting a bogus key or value into a '.keyval' file.
I still haven't got the implementation exactly right, however: with
this change, errors in .cfg files (for example) are reported twice.
Since the filename of the .keyval file is itself a Pyre property, this
creates order-of-initialization problems which exceed the limitations
of the Pyre framework.  I have one or two ideas about how to fix this,
but they are still in the oven, only half-baked...