Open output files once at the beginning of the run, and close them once at the end (instead of repeatedly opening and appending to them). This change elminates the need for the CIT_FC_OPEN_APPEND Autoconf test and the FORTRAN_POSITION_APPEND and FORTRAN_ACCESS_APPEND preprocessor macros.  It also (quite incidentally) made the Fortran routine viscos_setup() redundant.
Also, I fixed a bug I introduced.  In the Pyrex call to scan_timdat(),
I wrote "len(time_units)" instead of "sizeof(time_units)".  D'oh!
Mostly, the Pyrex calls are a transcription/merging of the old C++
bindings into the Python code, so hopefully this is the only bug of
its kind.