Add residual data member to Problem. Use setSolution() to set constraints. Remove _globalVec from Field.
Problem methods take global PETSc Vecs as arguments. The Problem methods
are responsible for managing the local vectors and putting the data back
in the global Vec.

Remove default argument from Field scatter methods and remove global Vec
data member. Can use "global" context for scatters if we need to hold a
global vector in Field. This makes the intent of the code clearer
overall.