Update for more flexible DMProjectFieldLocal().
No longer use superDM for updating state variables now that
DMProjectFieldLocal() can handle more complicated layouts.

We still have to put the updated state variables in their own vector so
that we have a copy of the current state variables and updated state
variables. At the end of the update we copy the updated state variables
into the auxiliary field.